refactor red tree iterators into own module and forward more features of the underlying iters
