+++ tests/it/main.rs
@@ -1,7 +1,7 @@
 mod basic;
 mod regressions;
 mod sendsync;
-#[cfg(feature = "serde1")]
+#[cfg(feature = "serialize")]
 mod serde;
 
 use cstree::{GreenNode, GreenNodeBuilder, Language, NodeCache, SyntaxKind};
